first draft of the new architecture to support subclasses and anonymous classes. A lot of tests are still failing
