+++ src/main/java/com/github/mauricioaniche/ck/metric/CBO.java
@@ -157,11 +157,6 @@ public class CBO extends ASTVisitor implements ClassLevelMetric, MethodLevelMetr
 	}
 
 	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
-	}
-
-	@Override
 	public void setResult(CKClassResult result) {
 		result.setCbo(getValue());
 	}
