+++ src/main/java/com/github/mauricioaniche/ck/metric/ClassLevelMetric.java
@@ -4,6 +4,5 @@ import com.github.mauricioaniche.ck.CKClassResult;
 import org.eclipse.jdt.core.dom.CompilationUnit;
 
 public interface ClassLevelMetric {
-	void execute(CompilationUnit cu, CKClassResult result);
 	void setResult(CKClassResult result);
 }
