+++ src/main/java/com/github/mauricioaniche/ck/metric/DIT.java
@@ -2,7 +2,6 @@ package com.github.mauricioaniche.ck.metric;
 
 import com.github.mauricioaniche.ck.CKClassResult;
 import org.eclipse.jdt.core.dom.ASTVisitor;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.ITypeBinding;
 import org.eclipse.jdt.core.dom.TypeDeclaration;
 
@@ -28,11 +27,6 @@ public class DIT extends ASTVisitor implements ClassLevelMetric {
 			calculate(father);
 		}
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
