+++ src/main/java/com/github/mauricioaniche/ck/metric/MethodLevelMetric.java
@@ -4,6 +4,5 @@ import com.github.mauricioaniche.ck.CKMethodResult;
 import org.eclipse.jdt.core.dom.MethodDeclaration;
 
 public interface MethodLevelMetric {
-	boolean visit(MethodDeclaration node);
 	void setResult(CKMethodResult result);
 }
