+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfAssignments.java
@@ -4,7 +4,6 @@ import com.github.mauricioaniche.ck.CKClassResult;
 import com.github.mauricioaniche.ck.CKMethodResult;
 import org.eclipse.jdt.core.dom.ASTVisitor;
 import org.eclipse.jdt.core.dom.Assignment;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.VariableDeclarationFragment;
 
 public class NumberOfAssignments extends ASTVisitor implements ClassLevelMetric, MethodLevelMetric {
@@ -29,11 +28,6 @@ public class NumberOfAssignments extends ASTVisitor implements ClassLevelMetric,
 	public void setResult(CKMethodResult result) {
 		result.setAssignmentsQty(qty);
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
