+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfMethods.java
@@ -2,7 +2,6 @@ package com.github.mauricioaniche.ck.metric;
 
 import com.github.mauricioaniche.ck.CKClassResult;
 import org.eclipse.jdt.core.dom.ASTVisitor;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.MethodDeclaration;
 import org.eclipse.jdt.core.dom.Modifier;
 
@@ -56,11 +55,6 @@ public class NumberOfMethods extends ASTVisitor implements ClassLevelMetric {
 			synchronizedMethods++;
 
 		return false;
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
