+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfNumbers.java
@@ -3,7 +3,6 @@ package com.github.mauricioaniche.ck.metric;
 import com.github.mauricioaniche.ck.CKClassResult;
 import com.github.mauricioaniche.ck.CKMethodResult;
 import org.eclipse.jdt.core.dom.ASTVisitor;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.NumberLiteral;
 
 public class NumberOfNumbers extends ASTVisitor implements ClassLevelMetric, MethodLevelMetric {
@@ -19,11 +18,6 @@ public class NumberOfNumbers extends ASTVisitor implements ClassLevelMetric, Met
 	public void setResult(CKMethodResult result) {
 		result.setNumbersQty(qty);
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
