+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfParenthesis.java
@@ -3,7 +3,6 @@ package com.github.mauricioaniche.ck.metric;
 import com.github.mauricioaniche.ck.CKClassResult;
 import com.github.mauricioaniche.ck.CKMethodResult;
 import org.eclipse.jdt.core.dom.ASTVisitor;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.ParenthesizedExpression;
 
 public class NumberOfParenthesis extends ASTVisitor implements ClassLevelMetric, MethodLevelMetric {
@@ -19,11 +18,6 @@ public class NumberOfParenthesis extends ASTVisitor implements ClassLevelMetric,
 	public void setResult(CKMethodResult result) {
 		result.setParenthesizedExpsQty(qty);
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
