+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfStringLiterals.java
@@ -2,7 +2,8 @@ package com.github.mauricioaniche.ck.metric;
 
 import com.github.mauricioaniche.ck.CKClassResult;
 import com.github.mauricioaniche.ck.CKMethodResult;
-import org.eclipse.jdt.core.dom.*;
+import org.eclipse.jdt.core.dom.ASTVisitor;
+import org.eclipse.jdt.core.dom.StringLiteral;
 
 public class NumberOfStringLiterals extends ASTVisitor implements ClassLevelMetric, MethodLevelMetric {
 
@@ -16,11 +17,6 @@ public class NumberOfStringLiterals extends ASTVisitor implements ClassLevelMetr
 	public void setResult(CKMethodResult result) {
 		result.setStringLiteralsQty(qty);
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
