+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfTryCatches.java
@@ -2,7 +2,8 @@ package com.github.mauricioaniche.ck.metric;
 
 import com.github.mauricioaniche.ck.CKClassResult;
 import com.github.mauricioaniche.ck.CKMethodResult;
-import org.eclipse.jdt.core.dom.*;
+import org.eclipse.jdt.core.dom.ASTVisitor;
+import org.eclipse.jdt.core.dom.TryStatement;
 
 public class NumberOfTryCatches extends ASTVisitor implements ClassLevelMetric, MethodLevelMetric {
 
@@ -17,11 +18,6 @@ public class NumberOfTryCatches extends ASTVisitor implements ClassLevelMetric, 
 	public void setResult(CKMethodResult result) {
 		result.setTryCatchQty(qty);
 
-	}
-
-	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
 	}
 
 	@Override
