+++ src/main/java/com/github/mauricioaniche/ck/metric/WMC.java
@@ -88,11 +88,6 @@ public class WMC extends ASTVisitor implements ClassLevelMetric, MethodLevelMetr
     }
 
 	@Override
-	public void execute(CompilationUnit cu, CKClassResult number) {
-		cu.accept(new IgnoreSubClasses(this));
-	}
-
-	@Override
 	public void setResult(CKClassResult result) {
 		result.setWmc(cc);
 	}
